refactor: move formInput element to components folder so that the reusable inputs can be used in login and register form
